+++ firmware/Core/Src/app_threadx.c
@@ -59,7 +59,7 @@ float                   g_vehicleSpeed;
 float 					g_currentSpeed;
 int16_t 				g_currentPWM;
 MotorControlState		g_motorControlState;
-float					g_targetSpeed;
+uint16_t				g_targetSpeed;
 unsigned char			trace_buffer[TRACE_BUFFER_SIZE];
 /* USER CODE END PV */
